refactor: split tools into submodules
