+++ src/lib.rs
@@ -1,7 +1,6 @@
 mod agent;
 mod api;
 mod config;
-mod errors;
 mod tools;
 
 pub use crate::agent::Agent;
